For moving procs to expired set, enqueue to original level
Previously, when a proc needs to be enqueued to the expired set
it is enqueued to RSDL_STARTING_LEVEL (i.e. the "original" level
when the proc was first enqueued). This was done to match how
I understood the linux rsdl patch (analog to static_prio)
https://web.archive.org/web/20070317221320/http://ck.kolivas.org/patches/staircase-deadline/2.6.21-rc3-mm2-rsdl-0.29.patch

This new behavior aims to match the behavior in the Project
Documentation, but I'm still not sure if my interpretation is correct.

find_vacant_queue() now takes in a second parameter which is the
level we want to put a proc in case it needs to be moved to
the expired set.