Do not re-prioritize sleeping processes
When proc gives up control to scheduler(), this could be
because it went to sleep, hence it should retain its position in queue

In any case, we only re-prioritize proc if it is still ready to run,
i.e. if p->state == RUNNABLE.

Thank you to @YSHebron for the bug report!