After preemption to scheduler, check if proc with higher priority exists
Since not ready (i.e. SLEEPing) processes stay in their positions,
It can be the case that the index of the current process being ran > 0
(e.g. if there are 2 sleeping procs in the queue, the proc at index 2
will be ran). However when such sleeping proc wakes up, it must be
chosen next by the scheduler instead of the process directly after
the proc that just finished (the proc that gave control back to scheduler)

We use the slow approach of manually checking each queue for now.
Since we only have 1 queue for Phase 1, this means that we check
the procs right BEFORE the process that just ran (ni < i) and check
if any of them are ready (i.e. RUNNABLE)