Add include guard to spinlock.h
For Phase 4, we will need to include it in both proc.h and proc.c
The include guard avoids redefinition of struct spinlock.

We can simply just include it in proc.h, which in turn is included in
proc.c, but proc.c has direct references to struct spinlock so I think
it's better to make this dependence explicit via includes.