Do not run procs in levels with depleted quantum
find_vacant_queue() is also renamed to find_available_queue()
to better match its behavior

We only acquire level queue lock AFTER checking that
level-local quantum is non-zero since q->ticks_left
is ONLY modified during swapping of sets which requires
that ptable.lock (which we are holding) is held.
Hence, we can be sure that no level-local quantum will be
replenished while we are searching for next proc to run.