Implement priofork syscall (NEEDS TESTING)
The allocproc() and fork() syscalls are extended to create
prioallocproc() and priofork(), respectively. The original
syscalls are now the default case of this new calls, indicated
by the default_value of -1.

TODO: verify is this default level will ONLY be used when swapping sets
(i.e. for leftover procs in the old active set), or also when proc
moves from active to expired set (e.g. moving to lower prio level
but none are available in the active set, hence go to expired set).

The former case is implemented in this commit.