When proc gives control back to scheduler, always re-enqueue
Section 2.3 says that a proc is re-enqueued to the next lower
prio level if it depleted its quantum, while it is re-enequeued
to the same level if it still has remaining quantum.

In either case, the process is re-enqueued. This is regardless
of the state of the proc after going back to the scheduler,
for example if it went to SLEEP and depleted its quantum at
the same time, it still needs to be re-enqueued to the next level.
The same is true for when the process is killed (state is ZOMBIE or UNUSED)