Fix schedlog not deactivated when there are no RUNNABLE procs
The check for disabling schedlog output (ticks > schedlog_lasttick)
was placed inside the if (found) clause in the scheduler, hence
schedlog_active = 0 was only done if there was a RUNNABLE proc found.

We move the check to the outermost loop in scheduler()

Thank you to @YSHebron for the detailed bug report!