Scheduler(): Only run procs that have positive quantum
Otherwise, there will be a short period of time
between the scheduling of the proc with depleted
quantum and the timer interrupt leading to a yield().
In this short period, the proc might be able to do
a little amount of work (even though it shouldn't have)

Thank you to @YSHebron ulit for finding the bug and for the fix!